pdf: Fix page index and
The module was not able to load any PDF with a single page (since
the index starts from 0, not 1 as it was assumed).

Also, fix a CID where Coverity was very very right.

Fixes CID 1356608:

The operaton may have an undefined behavior or yield to an
unexpected result.

In poppler_load_image(int, int): A bit shift operation has
a shift amount which is too large or has a negative value.