Revert "elementary: Fix issue of toggling border state causing visual jumps"
This reverts commit fb24955782b9242beb9447ebb6631b2b63921988.

While this patch did fix the issue of toggling window Borderless state
in EFl-Wl client applications, it is "Enlightenment's" stance that it
will not change to match this fix, so this has to be reverted else
internal E dialogs show up with no border (due to them using SSD).

This is going to re-open T4092 sadly :( and now leave that bug with no
possible fix.

ref T4092

Fixes T4126

Signed-off-by: Chris Michael <cp.michael@samsung.com>