evas/gl_common: (GL thread) Changed dummy surface creating routine from window to pbuffer
For EvasGL thread rendering, we can't use 'direct surface' which was same as Evas window surface.
Worker thread needs just a dummy surface for make current, but the worker thread can't use window surface
because the main thread uses it for rendering simultaneously.

On TIZEN wayland, There is an unknown error (segfault in wayland) creating additional native window,
therefore we can't create additional window surface for dummy purpose.
So we changed this routine to create a pbuffer instead it.

But after that, some backend (like gl_drm) has an side effect to create a pbuffer.
(these backends haven't implemented pbuffer functions)
So previous code (diret surface) is also affect for these cases,
Thread rendering seems to be hard to support these backends. (currently not)

Perhaps EGL_KHR_surfaceless_context can be final answer. (but it is an extension)

Change-Id: I8e2d9317bb533e897d24c5e0be4763040206d57d